Change cdef_index type to int8_t
cdef_bits is a literal of 2 bits and can have a maximum value of 3.
So we read literals with a maximum of 3 bits into cdef_index. And we
use -1 as a place holder. For these uses, int8_t will suffice.

PiperOrigin-RevId: 374499067

+++ src/frame_scratch_buffer.h
@@ -48,7 +48,7 @@ using IntraPredictionBuffer =
 // DecoderImpl::DecodeTiles().
 struct FrameScratchBuffer {
   LoopRestorationInfo loop_restoration_info;
-  Array2D<int16_t> cdef_index;
+  Array2D<int8_t> cdef_index;
   Array2D<TransformSize> inter_transform_sizes;
   BlockParametersHolder block_parameters_holder;
   TemporalMotionField motion_field;
